Removed Reservoir, merged CageTemplate into Cage and other related changes
