+++ refactor/config_data/test.json
@@ -15,23 +15,6 @@
         "description": "Sealice per day per cage/farm arriving from wildlife. Note that this value has been determined experimentally in the original code.",
         "value": 150
     },
-    "reservoir": {
-        "description": "Test reservoir",
-        "value": {
-            "num_fish": {
-                "description": "Starting number of fish",
-                "value": 4000
-            },
-            "enfish_res": {
-                "description": "Number of fish in millions in given month",
-                "value": [3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3]
-            },
-            "ewt_res": {
-                "description": "Average fish weight in kg",
-                "value": 2500
-            }
-        }
-    },
     "farms": [
         {
             "description": "Test farm",
