+++ refactor/tests/conftest.py
@@ -4,7 +4,6 @@ import pytest
 
 from src.Config import Config, to_dt
 from src.Farm import Farm
-from src.Reservoir import Reservoir
 
 @pytest.fixture
 def farm_config():
@@ -15,14 +14,6 @@ def farm_config():
 @pytest.fixture
 def farm(farm_config):
     return Farm(0, farm_config)
-
-
-@pytest.fixture
-def reservoir(farm_config):
-    total = 1000
-    farm_config.reservoir_num_lice = total
-    return Reservoir(farm_config)
-
 
 @pytest.fixture
 def first_cage(farm):
