+++ refactor/tests/test_Cage.py
@@ -129,3 +129,23 @@ class TestCage:
 
         _, lice_death = first_cage.update_fish_growth(1, 1)
         # assert lice_death == 0
+
+    def test_to_csv(self, first_cage):
+        first_cage.lice_population = {"L1": 1,
+                                      "L2": 2,
+                                      "L3": 3,
+                                      "L4": 4,
+                                      "L5f": 5,
+                                      "L5m": 6}
+        first_cage.num_fish = 7
+        first_cage.id = 0
+
+        csv_str = first_cage.to_csv()
+        csv_list = csv_str.split(", ")
+        print(csv_list)
+
+        assert csv_list[0] == "0"
+        assert csv_list[1] == "7"
+        for i in range(2, 7):
+            assert csv_list[i] == str(i - 1)
+        assert csv_list[8] == "21"
